Refactor: draggable card prop drilling 줄임
