+++ src/hooks/useWishList.js
@@ -8,6 +8,7 @@ export default function useWishList(item) {
   // TODO: card와 cardDetail 상태 공유. card에서 찜 누르면 cardDetail에서도 즉각 반영되도록
 
   useEffect(() => {
+    if (!item) return;
     setWish(item.wishListRestaurant);
   }, [item]);
